object-detection-docker/requirements.txt
requests==2.22.0
confluent_kafka==1.3.0
pyspark==2.4.5
Pillow==9.1.0
numpy